Largevalue should be double

+++ src/StdFace/StdFace_vals.h
@@ -86,5 +86,5 @@ int nintr;
 int **intrindx;
 double *intr;
 
-int LargeValue;
+double LargeValue;
 int S2;
